rename: customer domain package 변경
