+++ jpastudy/src/main/java/com/programmers/june/jpastudy/domain/customer/repository/CustomerRepository.java
@@ -1,5 +1,6 @@
-package com.programmers.june.jpastudy.domain.customer;
+package com.programmers.june.jpastudy.domain.customer.repository;
 
+import com.programmers.june.jpastudy.domain.customer.entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerRepository extends JpaRepository<Customer, Long> {
